Fix Bug couldn't delete locations with no photos.

+++ quite_study_spaces_app/lib/states/user_profile_state.dart
@@ -119,9 +119,11 @@ class UserProfileState extends ChangeNotifier{
         final data = locationDoc.data();
         final photoURL = data?['photoURL'] ?? '';
           final file = File(photoURL);
-          if(await file.exists()) {
-            await file.delete();
-            print("Local image deleted: $photoURL");
+          if(photoURL.isNotEmpty){
+            if(await file.exists()) {
+              await file.delete();
+              print("Local image deleted: $photoURL");
+            }
           }
       }
       await FirebaseFirestore.instance.collection('users').doc(user.uid).update({
